2.3.1 solve!

+++ src/main/java/web/dao/UserDAO.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface UserDAO {
     List<User> getUsers();
     void createUser(User user);
-    User readUser(int id);
-    void updateUser(int id, User user);
-    void deleteUser(int id);
+    User readUser(Long id);
+    void updateUser(Long id, User user);
+    void deleteUser(Long id);
 }
