+++ src/main/java/web/dao/UserDAOImpl.java
@@ -1,11 +1,11 @@
 package web.dao;
 
-import jakarta.persistence.EntityManager;
-import jakarta.persistence.PersistenceContext;
 import org.springframework.stereotype.Repository;
 import org.springframework.transaction.annotation.Transactional;
 import web.model.User;
 
+import javax.persistence.EntityManager;
+import javax.persistence.PersistenceContext;
 import java.util.List;
 
 @Repository
@@ -28,19 +28,19 @@ public class UserDAOImpl implements UserDAO {
 
     @Override
     @Transactional
-    public User readUser(int id) {
+    public User readUser(Long id) {
         return entityManager.find(User.class, id);
     }
 
     @Override
     @Transactional
-    public void updateUser(int id, User user) {
+    public void updateUser(Long id, User user) {
         entityManager.merge(user);
     }
 
     @Override
     @Transactional
-    public void deleteUser(int id) {
+    public void deleteUser(Long id) {
         User user = entityManager.find(User.class, id);
         entityManager.remove(user);
     }
