+++ src/main/java/web/service/UserService.java
@@ -7,7 +7,7 @@ import java.util.List;
 public interface UserService {
     List<User> getUsers();
     void createUser(User user);
-    User readUser(int id);
-    void deleteUser(int id);
-    void updateUser(int id, User user);
+    User readUser(Long id);
+    void deleteUser(Long id);
+    void updateUser(Long id, User user);
 }
